FlappyCode/src/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
Werkzeug==3.1.3 
gunicorn==23.0.0
python-logging-loki==0.3.1
pyodbc==5.2.0 
python-dotenv==1.0.1 
azure-keyvault-secrets==4.9.0
azure-identity==1.19.0
